+++ src/orcAI/__init__.py
@@ -1,6 +1,6 @@
 import os
 
-__version__ = "1.1.10dev1"
+__version__ = "1.1.10"
 
 os.environ["KERAS_BACKEND"] = "tensorflow"
 os.environ["TF_CPP_MIN_LOG_LEVEL"] = "3"